fix: move scroll-to-top button to fixed position outside sidebar
Extract scroll-to-top from SideBar into a standalone fixed-position
component (like Bluesky's LoadLatestBtn). The button floats at
bottom-left of the viewport without affecting sidebar layout. Only
renders on desktop when sidebar is visible.
