+++ packages/frontend/app/(app)/_layout.tsx
@@ -10,6 +10,7 @@ import KeyboardShortcutsModal from "@/components/KeyboardShortcutsModal";
 import RegisterPush from '@/components/RegisterPushToken';
 import { RealtimePostsBridge } from '@/components/RealtimePostsBridge';
 import { RightBar } from "@/components/RightBar";
+import { ScrollToTopButton } from "@/components/ScrollToTopButton";
 import { SideBar } from "@/components/SideBar";
 import { SignInBanner } from "@/components/SignInBanner";
 import { ThemedView } from "@/components/ThemedView";
@@ -111,6 +112,7 @@ export default function AppLayout() {
       <MainLayout isScreenNotMobile={isScreenNotMobile} />
       <RegisterPush />
       {isAuthenticated && !isScreenNotMobile && !keyboardVisible && <BottomBar />}
+      {isScreenNotMobile && <ScrollToTopButton />}
       {!isScreenNotMobile && <DrawerOverlay />}
       {!isAuthenticated && <SignInBanner />}
       <WelcomeModalGate appIsReady={true} />
